tehtävä 4.21

+++ 4/blogilista/controllers/blogs.js
@@ -36,8 +36,19 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+  console.log(blog.user.toString())
+  console.log(decodedToken.id.toString())
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'deletion is unauthorized' })
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
